Move observation results payloads to separate file
`ObservationsController.kt` is getting unreasonably large, mostly due to the
various payload class definitions. The biomass payloads have already been split
out into their own source file; do the same for the monitoring results payloads.

No code changes here, just moving classes to a different file in the same package.
